refactor: extract the helper method to a separate file
