+++ Samples~/AvatarCreatorSamples/AvatarCreatorWizard/Scripts/UI/SelectionScreens/AvatarCreatorSelection.cs
@@ -202,24 +202,12 @@ namespace ReadyPlayerMe.Samples.AvatarCreatorWizard
             var startTime = Time.time;
             var colors = await avatarManager.LoadAvatarColors();
             
-            var colorAssetTypes = GetAssetTypesByFilter(AssetFilter.Color).ToHashSet();
+            var colorAssetTypes = AssetFilterHelper.GetAssetTypesByFilter(AssetFilter.Color).ToHashSet();
             var equippedColors = AvatarCreatorData.AvatarProperties.Assets.Where(kvp => colorAssetTypes.Contains(kvp.Key))
                      .ToDictionary(kvp => kvp.Key, kvp => Convert.ToInt32(kvp.Value));
 
             assetButtonCreator.CreateColorUI(colors, UpdateAvatar, equippedColors);
             SDKLogger.Log(TAG, $"All colors loaded in {Time.time - startTime:F2}s");
-        }
-
-        public static IEnumerable<AssetType> GetAssetTypesByFilter(AssetFilter filter)
-        {
-            return Enum.GetValues(typeof(AssetType))
-                .Cast<AssetType>()
-                .Where(assetType =>
-                {
-                    var fieldInfo = typeof(AssetType).GetField(assetType.ToString());
-                    var attribute = fieldInfo?.GetCustomAttribute<AssetTypeFilterAttribute>();
-                    return attribute?.filter == filter;
-                });
         }
 
         private void CreateUI()
